Resizable windows, added back some more GUI Widgets and got the compiler to shut up on a few things

+++ Kernel/include/arch/x86_64/serial.h
@@ -4,5 +4,5 @@ void initialize_serial();
 void unlockSerial();
 void write_serial(const char c);
 void write_serial(const char* s);
-void write_serial_n(const char* s, unsigned long long n);
+void write_serial_n(const char* s, unsigned n);
 void write_serial_string(const char* s);
